# core: exploit the use of cache in txSortedMap