This PR reduces the filter overhead in `txList`. 
It could be optimized by exploiting the cache field, where transactions are always sorted by nonce in `txSortedMap`. And thus the lowest nonce could be easily accessed by taking the first element in `Filter` of `txList`.

The maintenance of heap and cache could be done alongside the filter loop without much memory overhead. 
Sorted cache generation is encapsulated in a method that could be reused in another two methods. 

The benchmark comparison is shown below. There is a little improvement through these tests. 
Though I have no idea why larger batch insert benchmark slows down a little bit. 
```
benchmark                           old ns/op      new ns/op      delta
BenchmarkPendingDemotion100-4       172            173            +0.58%
BenchmarkPendingDemotion1000-4      172            173            +0.58%
BenchmarkPendingDemotion10000-4     175            175            +0.00%
BenchmarkFuturePromotion100-4       1333           1335           +0.15%
BenchmarkFuturePromotion1000-4      1333           1334           +0.08%
BenchmarkFuturePromotion10000-4     1323           1330           +0.53%
BenchmarkPoolInsert-4               158419         158417         -0.00%
BenchmarkPoolBatchInsert100-4       15538644       15547263       +0.06%
BenchmarkPoolBatchInsert1000-4      157286911      157074763      -0.13%
BenchmarkPoolBatchInsert10000-4     1534317650     1533876689     -0.03%
```