@fjl 2018-09-03
The code looks a bit nicer, but your benchmark result doesn't seem like it's worth making this change.